WIP, cannibalism from KrampusHack2018

+++ src/Location.cpp
@@ -11,8 +11,8 @@ bool Location::ValidForMaze(Maze* m) {
    if (x < 1) {return false;}
    if (y < 1) {return false;}
    if (z < 1) {return false;}
-   if (x >= m->SIZE_EW) {return false;}
-   if (y >= m->SIZE_NS) {return false;}
-   if (z >= m->SIZE_UD) {return false;}
+   if (x >= (int)m->SIZE_EW) {return false;}
+   if (y >= (int)m->SIZE_NS) {return false;}
+   if (z >= (int)m->SIZE_UD) {return false;}
    return true;
 }
